+++ test/array/index.js
@@ -7,6 +7,7 @@ module.exports = {
     forEach: require('./forEach'),
     map: require('./map'),
     reduce: require('./reduce'),
+    reduceRight: require('./reduceRight'),
     reverse: require('./reverse'),
     some: require('./some')
 };